requirements.txt
APScheduler==3.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gspread==5.8.0
httplib2==0.22.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==1.0.0
python-telegram-bot==13.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tornado==6.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
